tools/requirements.txt
# YOLOv8 检测服务依赖

ultralytics>=8.0.0
flask>=2.0.0
pillow>=9.0.0
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
